# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Collectarr",
    "description": "Collectarr metadata and library backend API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/metadata/catalog/items": {
      "get": {
        "tags": [
          "metadata",
          "catalog-items"
        ],
        "summary": "Search Catalog Items",
        "operationId": "search_catalog_items_api_v1_metadata_catalog_items_get",
        "parameters": [
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "anime",
                    "boardgame",
                    "book",
                    "comic",
                    "game",
                    "manga",
                    "movie",
                    "music",
                    "tv"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "identifier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Identifier"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CatalogItemSummaryV1"
                  },
                  "title": "Response Search Catalog Items Api V1 Metadata Catalog Items Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "metadata",
          "catalog-items"
        ],
        "summary": "Create Catalog Item",
        "operationId": "create_catalog_item_api_v1_metadata_catalog_items_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogItemWriteV1"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemV1"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/metadata/catalog/items/{item_id}": {
      "get": {
        "tags": [
          "metadata",
          "catalog-items"
        ],
        "summary": "Get Catalog Item",
        "operationId": "get_catalog_item_api_v1_metadata_catalog_items__item_id__get",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemV1"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "metadata",
          "catalog-items"
        ],
        "summary": "Update Catalog Item",
        "operationId": "update_catalog_item_api_v1_metadata_catalog_items__item_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Item Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CatalogItemWriteV1"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CatalogItemV1"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnimeCatalogDetailsV1": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "sort_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Title"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "release_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDateValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "identifiers": {
            "items": {
              "$ref": "#/components/schemas/CatalogIdentifierV1"
            },
            "type": "array",
            "title": "Identifiers"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/CatalogImageV1"
            },
            "type": "array",
            "title": "Images"
          },
          "kind": {
            "type": "string",
            "const": "anime",
            "title": "Kind"
          },
          "studios": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Studios"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          },
          "credits": {
            "items": {
              "$ref": "#/components/schemas/CatalogCreditV1"
            },
            "type": "array",
            "title": "Credits"
          },
          "format": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "region": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Region"
          },
          "edition_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition Title"
          },
          "episodes": {
            "items": {
              "$ref": "#/components/schemas/CatalogEpisodeV1"
            },
            "type": "array",
            "title": "Episodes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "kind"
        ],
        "title": "AnimeCatalogDetailsV1"
      },
      "BoardGameCatalogDetailsV1": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "sort_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Title"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "release_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDateValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "identifiers": {
            "items": {
              "$ref": "#/components/schemas/CatalogIdentifierV1"
            },
            "type": "array",
            "title": "Identifiers"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/CatalogImageV1"
            },
            "type": "array",
            "title": "Images"
          },
          "kind": {
            "type": "string",
            "const": "boardgame",
            "title": "Kind"
          },
          "edition": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition"
          },
          "publishers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Publishers"
          },
          "designers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Designers"
          },
          "mechanics": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Mechanics"
          },
          "categories": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Categories"
          },
          "min_players": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Players"
          },
          "max_players": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Players"
          },
          "recommended_players": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Recommended Players"
          },
          "best_players": {
            "items": {
              "type": "integer"
            },
            "type": "array",
            "maxItems": 32,
            "title": "Best Players"
          },
          "play_time_minutes": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Play Time Minutes"
          },
          "components": {
            "items": {
              "$ref": "#/components/schemas/CatalogComponentV1"
            },
            "type": "array",
            "title": "Components"
          },
          "related_items": {
            "items": {
              "$ref": "#/components/schemas/CatalogRelatedItemV1"
            },
            "type": "array",
            "title": "Related Items"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "kind"
        ],
        "title": "BoardGameCatalogDetailsV1"
      },
...
      "BookCatalogDetailsV1": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "sort_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Title"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "release_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDateValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "identifiers": {
            "items": {
              "$ref": "#/components/schemas/CatalogIdentifierV1"
            },
            "type": "array",
            "title": "Identifiers"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/CatalogImageV1"
            },
            "type": "array",
            "title": "Images"
          },
          "kind": {
            "type": "string",
            "const": "book",
            "title": "Kind"
          },
          "contributors": {
            "items": {
              "$ref": "#/components/schemas/CatalogCreditV1"
            },
            "type": "array",
            "title": "Contributors"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "publication_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDateValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "edition": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Edition"
          },
          "format": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "genres": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Genres"
          },
          "subjects": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Subjects"
          },
          "original_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Title"
          },
          "series_membership": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatalogSeriesMembershipV1"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "kind"
        ],
        "title": "BookCatalogDetailsV1"
      },
...
      "CatalogComponentV1": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "quantity": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Quantity",
            "default": 1
          },
          "position": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Position",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CatalogComponentV1"
      },
...
      "CatalogIdentifierV1": {
        "properties": {
          "identifier_type": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Identifier Type"
          },
          "value": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Value"
          },
          "is_primary": {
            "type": "boolean",
            "title": "Is Primary",
            "default": false
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "identifier_type",
          "value"
        ],
        "title": "CatalogIdentifierV1"
      },
      "CatalogImageV1": {
        "properties": {
          "image_type": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Image Type"
          },
          "url": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2083,
                "minLength": 1,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "image_key": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 512
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Key"
          },
          "position": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Position",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "image_type"
        ],
        "title": "CatalogImageV1"
      },
      "CatalogItemDetailsV1": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/AnimeCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/BoardGameCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/BookCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/ComicCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/GameCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/MangaCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/MovieCatalogDetailsV1-Output"
          },
          {
            "$ref": "#/components/schemas/MusicCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/TVCatalogDetailsV1"
          }
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "anime": "#/components/schemas/AnimeCatalogDetailsV1",
            "boardgame": "#/components/schemas/BoardGameCatalogDetailsV1",
            "book": "#/components/schemas/BookCatalogDetailsV1",
            "comic": "#/components/schemas/ComicCatalogDetailsV1",
            "game": "#/components/schemas/GameCatalogDetailsV1",
            "manga": "#/components/schemas/MangaCatalogDetailsV1",
            "movie": "#/components/schemas/MovieCatalogDetailsV1-Output",
            "music": "#/components/schemas/MusicCatalogDetailsV1",
            "tv": "#/components/schemas/TVCatalogDetailsV1"
          }
        }
      },
      "CatalogItemSummaryV1": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "enum": [
              "anime",
              "boardgame",
              "book",
              "comic",
              "game",
              "manga",
              "movie",
              "music",
              "tv"
            ],
            "title": "Kind"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "sort_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Title"
          },
          "release_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDateValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "cover_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cover Image Url"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "format": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Format"
          },
          "country": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Country"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "barcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "kind",
          "title"
        ],
        "title": "CatalogItemSummaryV1"
      },
      "CatalogItemV1": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "details": {
            "$ref": "#/components/schemas/CatalogItemDetailsV1"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "details",
          "created_at",
          "updated_at"
        ],
        "title": "CatalogItemV1"
      },
      "CatalogItemWriteDetailsV1": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/AnimeCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/BoardGameCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/BookCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/ComicCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/GameCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/MangaCatalogDetailsV1"
          },
          {
            "$ref": "#/components/schemas/MovieCatalogDetailsV1-Input"
          },
          {
            "$ref": "#/components/schemas/MusicCatalogWriteDetailsV1"
          },
          {
            "$ref": "#/components/schemas/TVCatalogDetailsV1"
          }
        ],
        "discriminator": {
          "propertyName": "kind",
          "mapping": {
            "anime": "#/components/schemas/AnimeCatalogDetailsV1",
            "boardgame": "#/components/schemas/BoardGameCatalogDetailsV1",
            "book": "#/components/schemas/BookCatalogDetailsV1",
            "comic": "#/components/schemas/ComicCatalogDetailsV1",
            "game": "#/components/schemas/GameCatalogDetailsV1",
            "manga": "#/components/schemas/MangaCatalogDetailsV1",
            "movie": "#/components/schemas/MovieCatalogDetailsV1-Input",
            "music": "#/components/schemas/MusicCatalogWriteDetailsV1",
            "tv": "#/components/schemas/TVCatalogDetailsV1"
          }
        }
      },
      "CatalogItemWriteV1": {
        "properties": {
          "details": {
            "$ref": "#/components/schemas/CatalogItemWriteDetailsV1"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "details"
        ],
        "title": "CatalogItemWriteV1"
      },
...
      "ComicCatalogDetailsV1": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 500,
            "minLength": 1,
            "title": "Title"
          },
          "sort_title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Title"
          },
          "subtitle": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Subtitle"
          },
          "release_date": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PartialDateValue"
              },
              {
                "type": "null"
              }
            ]
          },
          "identifiers": {
            "items": {
              "$ref": "#/components/schemas/CatalogIdentifierV1"
            },
            "type": "array",
            "title": "Identifiers"
          },
          "images": {
            "items": {
              "$ref": "#/components/schemas/CatalogImageV1"
            },
            "type": "array",
            "title": "Images"
          },
          "kind": {
            "type": "string",
            "const": "comic",
            "title": "Kind"
          },
          "series": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CatalogSeriesMembershipV1"
              },
              {
                "type": "null"
              }
            ]
          },
          "issue_number": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Issue Number"
          },
          "variant": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Variant"
          },
          "publisher": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255
              },
              {
                "type": "null"
              }
            ],
            "title": "Publisher"
          },
          "plot": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Plot"
          },
          "creators": {
            "items": {
              "$ref": "#/components/schemas/CatalogCreditV1"
            },
            "type": "array",
            "title": "Creators"
          },
          "characters": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Characters"
          },
          "key_issue": {
            "type": "boolean",
            "title": "Key Issue",
            "default": false
          },
          "story_arcs": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Story Arcs"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "kind"
        ],
        "title": "ComicCatalogDetailsV1"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PartialDateValue": {
        "properties": {
          "year": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 9999.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "month": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 12.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Month"
          },
          "day": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 31.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Day"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "PartialDateValue",
        "description": "A catalog date with independently optional components."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}